+++ packages/atomic/src/components/atomic-result-list/atomic-result-list.tsx
@@ -29,7 +29,7 @@ export class AtomicResultList {
    * Css class for a list element
    */
   @Prop() listElementClass = '';
-
+  @Prop() fieldsToInclude = '';
   @Element() host!: HTMLDivElement;
   @State() state!: ResultListState;
 
@@ -38,12 +38,19 @@ export class AtomicResultList {
   private resultList!: ResultList;
   private resultTemplatesManager!: ResultTemplatesManager<string>;
 
+  private get fields() {
+    if (this.fieldsToInclude.trim() === '') return;
+    return this.fieldsToInclude.split(',').map((field) => field.trim());
+  }
+
   @Initialization()
   public initialize() {
     this.resultTemplatesManager = new ResultTemplatesManager<string>(
       this.engine
     );
-    this.resultList = buildResultList(this.engine);
+    this.resultList = buildResultList(this.engine, {
+      options: {fieldsToInclude: this.fields},
+    });
     this.unsubscribe = this.resultList.subscribe(() => this.updateState());
 
     this.registerDefaultResultTemplates();