+++ packages/headless/src/features/redirection/redirection-slice.test.ts
@@ -13,8 +13,12 @@ import {
   MockEngine,
 } from '../../test';
 import {SearchAppState} from '../../state/search-app-state';
+import {NoopPreprocessRequestMiddleware} from '../../api/platform-client';
+import pino from 'pino';
 
 describe('redirection slice', () => {
+  const logger = pino({level: 'silent'});
+
   it('should have initial state', () => {
     expect(redirectionReducer(undefined, {type: 'randomAction'})).toEqual(
       getRedirectionInitialState()
@@ -51,7 +55,11 @@ describe('redirection slice', () => {
 
   let engine: MockEngine<SearchAppState>;
   async function mockPlan(trigger?: Trigger) {
-    const apiClient = new SearchAPIClient(async () => '');
+    const apiClient = new SearchAPIClient({
+      renewAccessToken: async () => '',
+      logger,
+      preprocessRequest: NoopPreprocessRequestMiddleware,
+    });
     const triggers = trigger ? [trigger] : [];
     jest.spyOn(apiClient, 'plan').mockResolvedValue({
       success: {
@@ -67,6 +75,7 @@ describe('redirection slice', () => {
     })(engine.dispatch, () => createMockState(), {
       searchAPIClient: apiClient,
       analyticsClientMiddleware: (_, p) => p,
+      logger,
     });
 
     return response;